requirements.txt
transformers
tokenizers
diffusers
tqdm
pillow
gradio
torch
torchvision
fastapi